+++ cmd/qrterminal/main.go
@@ -16,7 +16,7 @@ import (
 var verboseFlag bool
 var levelFlag string
 var quietZoneFlag int
-var sixelDisableFlag bool
+var sixelFlag bool
 
 func getLevel(s string) qr.Level {
 	switch l := strings.ToLower(s); l {
@@ -35,7 +35,7 @@ func main() {
 	flag.BoolVar(&verboseFlag, "v", false, "Output debugging information")
 	flag.StringVar(&levelFlag, "l", "L", "Error correction level")
 	flag.IntVar(&quietZoneFlag, "q", 2, "Size of quietzone border")
-	flag.BoolVar(&sixelDisableFlag, "s", false, "disable sixel format for output")
+	flag.BoolVar(&sixelFlag, "sixel", false, "enable sixel format for output (opt-in)")
 
 	flag.Parse()
 	level := getLevel(levelFlag)
@@ -62,7 +62,7 @@ func main() {
 		BlackChar: qrterminal.BLACK,
 		WhiteChar: qrterminal.WHITE,
 	}
-	if !sixelDisableFlag {
+	if sixelFlag {
 		cfg.WithSixel = qrterminal.IsSixelSupported(os.Stdout)
 	}
 	if verboseFlag {